feat(frontend): Optimize JNI to use tokio-uring (#263)
* Adapt to tokio-uring


* feat(frontend): Add getFrontend command & createStream command


* feat(frontend): Add openStream command


* feat(frontend): Add minOffset and maxOffset command


* feat(frontend): Add append and read command


* fix(frontend): Fix read argument to start_offset, end_offset


* fix(frontend): Change the return type of create_stream to u64


* fix(frontend): Change the java return type of create_stream to CompletableFuture<Long>


---------

+++ sdks/rust/src/stream.rs
@@ -72,7 +72,7 @@ impl Stream {
     pub async fn read(
         &self,
         start_offset: i64,
-        end_offset: i32,
+        end_offset: i64,
         batch_max_bytes: i32,
     ) -> Result<Bytes, ClientError> {
         trace!(
